API/requirements.txt
django~=5.2.3
pymongo~=4.13.2
djangorestframework~=3.16.0
django-cors-headers